migrate to bazel
