+++ .bazelrc
@@ -1,0 +1,2 @@
+build --nojava_header_compilation --java_language_version=21
+build --nojava_header_compilation --java_runtime_version=remotejdk_21
