--- src/test/java/org/example/cel/CelTestWithMessageBuilder.java
@@ -1,84 +1,0 @@
-package org.example.cel;
-
-import com.google.protobuf.Descriptors;
-import com.google.protobuf.DynamicMessage;
-import com.google.protobuf.InvalidProtocolBufferException;
-import com.google.protobuf.Message;
-import com.google.protobuf.util.JsonFormat;
-import dev.cel.common.CelValidationException;
-import dev.cel.common.types.CelType;
-import dev.cel.common.types.MapType;
-import dev.cel.common.types.SimpleType;
-import dev.cel.common.types.StructTypeReference;
-import dev.cel.runtime.CelEvaluationException;
-import dev.cel.runtime.CelRuntime;
-import org.example.utils.ProtoUtils;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public class CelTestWithMessageBuilder {
-    @Test
-    public void testReferDescriptorViaGeneratedClass() throws InvalidProtocolBufferException, CelValidationException, CelEvaluationException {
-        List<Descriptors.FileDescriptor> fileDescriptors= ProtoUtils.buildFileDescriptors("/test.pb");
-        String query = "contents.map_field_b.size() > 0";
-        Map<String, CelType> vars = Map.ofEntries(
-                Map.entry("map_field_a", MapType.create(SimpleType.STRING, SimpleType.STRING)),
-                Map.entry("contents", StructTypeReference.create("test.Test.Contents"))
-        );
-
-        CelRuntime.Program program = CelUtils.buildProgram(vars, query, org.example.proto.Test.getDescriptor());
-        String stringJson = """
-                {
-                    "contents" : {
-                        "map_field_b" : {
-                            "foo" : "bar"
-                        }
-                    }
-                }
-                """;
-        // =========
-        // Get Descriptor from Test class which generated by protoc
-        // =========
-        DynamicMessage.Builder messageBuilder = DynamicMessage.newBuilder(org.example.proto.Test.getDescriptor());
-        JsonFormat.parser().ignoringUnknownFields().merge(stringJson, messageBuilder);
-
-        // This will OK
-        Assertions.assertDoesNotThrow(() -> program.eval(toMap(messageBuilder.build())));
-    }
-
-    @Test
-    public void testUsingGeneratedMessageBuilder() throws InvalidProtocolBufferException, CelValidationException, CelEvaluationException {
-        List<Descriptors.FileDescriptor> fileDescriptors= ProtoUtils.buildFileDescriptors("/test.pb");
-        String query = "contents.map_field_b.size() > 0";
-        Map<String, CelType> vars = Map.ofEntries(
-                Map.entry("map_field_a", MapType.create(SimpleType.STRING, SimpleType.STRING)),
-                Map.entry("contents", StructTypeReference.create("test.Test.Contents"))
-        );
-
-        CelRuntime.Program program = CelUtils.buildProgram(vars, query, org.example.proto.Test.getDescriptor());
-        Message message = org.example.proto.Test.newBuilder()
-                .setContents(
-                        org.example.proto.Test.Contents.newBuilder()
-                                .putAllMapFieldB(
-                                        Map.of("foo", "bar")
-                                )
-                                .build()
-                ).build();
-
-        // This will OK
-        Assertions.assertDoesNotThrow(() -> program.eval(toMap(message)));
-    }
-
-    private static Map<String, ?> toMap(Message message) {
-        Map<String, Object> result = new HashMap<>();
-        for (Descriptors.FieldDescriptor d : message.getDescriptorForType().getFields()) {
-            Object fieldValue = message.getAllFields().getOrDefault(d, message.getField(d));
-            result.put(d.getName(), fieldValue);
-        }
-        return result;
-    }
-}
